refactor: update type db of values

+++ src/expense/expense.entity.ts
@@ -10,7 +10,8 @@ export class ExpenseEntity extends BaseEntity {
   description: string
 
   @Column({
-    type: 'double',
+    type: 'decimal',
+    precision: 6,
     scale: 2,
     nullable: false
   })
